Optic functors
- Basic tests in `test_optic.py`
- Add explicit dtypes to all classes
  - This actually sucks, so later I'll remove this + bake dtype into
    choice of FiniteFunction implementation.

+++ open_hypergraphs/array/numpy.py
@@ -124,4 +124,4 @@ class NumpyBackend(ArrayBackend):
 
         # compute & return connected components
         c, cc_ix = sparse.csgraph.connected_components(M)
-        return c, cc_ix
+        return c, cc_ix.astype(source.dtype)
